requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.39.0
pandas>=2.1.0
openpyxl>=3.1.0
python-docx>=1.0.0
pypdf>=3.17.0
python-dotenv>=1.0.0
numpy>=1.26.0
python-multipart>=0.0.6
aiofiles>=23.2.0
